Reactのsandbox開発環境としての整備: Router導入
