+++ src/pages/Home.tsx
@@ -1,0 +1,11 @@
+export const Home = () => {
+  return (
+    <>
+      <div className="home">
+        <div className="card">
+          <p>This is Home Component.</p>
+        </div>
+      </div>
+    </>
+  );
+};
